requirements.txt
streamlit
fpdf2
plotly
streamlit-option-menu
scikit-learn
PyPDF2
sentence-transformers
keybert
google-generativeai